requirements.txt
PyOpenGL 
PyOpenGL_accelerate 
glfw 
numpy 
pyrr
pygame